Created RESTFul Request class to create and execute requests.
The request is complete and executing. I am still to create response parsing functionality.

I have decided that this functionality should exist within the same class as it is currently only intended to work with this class.

I will need to extend this class to introduce authentication for secure Parplu/Bandu type requests.